Delete CNAME in Master Branch
CNAME is mainly used for custom domain with Github Pages. CNAME should
be put in the root of Master branch ONLY when the repository is a `User
Page Site` like `username.github.io`. Otherwise, it should be in the
root of `gh-pages` Branch.

This file may lead a misunderstanding of master branch for the freshers.